solana_curve25519

Module curve_syscall_traits

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

The traits representing the basic elliptic curve operations.

These traits are instantiatable by all the commonly used elliptic curves and should help in organizing syscall support for other curves in the future. more complicated or curve-specific functions that are needed in cryptographic applications should be representable by combining the associated functions of these traits.

NOTE: This module temporarily lives in zk_token_sdk/curve25519, but it is independent of zk-token-sdk or curve25519. It should be moved to a more general location in the future.

Constants

ADDDeprecated

CURVE25519_EDWARDSDeprecated

CURVE25519_RISTRETTODeprecated

MULDeprecated

SUBDeprecated

Traits

GroupOperationsDeprecated

MultiScalarMultiplicationDeprecated

PairingDeprecated

PointValidationDeprecated