solana_curve25519::curve_syscall_traits

Trait GroupOperations

pub trait GroupOperations {
    type Point;
    type Scalar;

    // Required methods
    fn add(
        left_point: &Self::Point,
        right_point: &Self::Point,
    ) -> Option<Self::Point>;
    fn subtract(
        left_point: &Self::Point,
        right_point: &Self::Point,
    ) -> Option<Self::Point>;
    fn multiply(
        scalar: &Self::Scalar,
        point: &Self::Point,
    ) -> Option<Self::Point>;
}

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Required Associated Types

type Point

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

type Scalar

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Required Methods

fn add( left_point: &Self::Point, right_point: &Self::Point, ) -> Option<Self::Point>

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Adds two curve points: P_0 + P_1.

fn subtract( left_point: &Self::Point, right_point: &Self::Point, ) -> Option<Self::Point>

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Subtracts two curve points: P_0 - P_1.

NOTE: Altneratively, one can consider replacing this with a negate function that maps a curve point P -> -P. Then subtraction can be computed by combining negate and add syscalls. However, subtract is a much more widely used function than negate.

fn multiply(scalar: &Self::Scalar, point: &Self::Point) -> Option<Self::Point>

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Multiplies a scalar S with a curve point P: S*P

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.