solana_curve25519::curve_syscall_traits

Trait PointValidation

pub trait PointValidation {
    type Point;

    // Required method
    fn validate_point(&self) -> bool;
}

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Required Associated Types

type Point

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Required Methods

fn validate_point(&self) -> bool

👎Deprecated since 3.1.0: This crate has been marked for formal inclusion in the Agave Unstable API. From v4.0.0 onward, the agave-unstable-api crate feature must be specified to acknowledge use of an interface that may break without warning.

Verifies if a byte representation of a curve point lies in the curve.