solana_curve25519
Module curve_syscall_traits

The traits representing the basic elliptic curve operations.

These traits are instantiatable by all the commonly used elliptic curves and should help in organizing syscall support for other curves in the future. more complicated or curve-specific functions that are needed in cryptographic applications should be representable by combining the associated functions of these traits.

NOTE: This module temporarily lives in zk_token_sdk/curve25519, but it is independent of zk-token-sdk or curve25519. It should be moved to a more general location in the future.

Constants

ADD

CURVE25519_EDWARDS

CURVE25519_RISTRETTO

MUL

SUB

Traits

GroupOperations

MultiScalarMultiplication

Pairing

PointValidation